Replace edit-portfolio item test file with the standalone version
